src/main/java/com/fiduprevisora/service/OrderService.java
package com.fiduprevisora.service;

import com.fiduprevisora.entity.Orders;
import com.fiduprevisora.repository.OrderRepository;
import jakarta.transaction.Transactional;
import lombok.AllArgsConstructor;
import org.springframework.stereotype.Service;

import java.util.List;

@Transactional
@AllArgsConstructor
@Service
public class OrderService {

    private OrderRepository orderRepository;

    public List<Orders> findAllOrders() {
        return orderRepository.findAll();
    }

    public Orders getOrderById(Long id) {
        return orderRepository.findById(id).orElse(null);
    }

    public Orders saveOrder(Orders order) {
        return orderRepository.save(order);
    }

    public void deleteOrder(Long id) {
        orderRepository.deleteById(id);
    }
}
